+++ src/Taxes/Calculators/DefaultTaxCalculator.php
@@ -17,6 +17,7 @@ namespace Vanilo\Taxes\Calculators;
 use Nette\Schema\Expect;
 use Nette\Schema\Schema;
 use Vanilo\Adjustments\Adjusters\SimpleTax;
+use Vanilo\Adjustments\Contracts\Adjustable;
 use Vanilo\Contracts\DetailedAmount;
 use Vanilo\Support\Dto\DetailedAmount as DetailedAmountDto;
 use Vanilo\Taxes\Contracts\TaxCalculator;
@@ -28,29 +29,44 @@ class DefaultTaxCalculator implements TaxCalculator
         return 'Default (rate based)';
     }
 
-    public function getAdjuster(?array $configuration = null): ?object
+    public function getAdjuster(?array $configuration = null): SimpleTax
     {
         $rate = floatval($configuration['rate'] ?? 0);
-        $adjuster = new SimpleTax($rate, false);
-        $adjuster->setTitle("$rate%");
+        $adjuster = new SimpleTax($rate, (bool) ($configuration['included'] ?? false));
+        $adjuster->setTitle($configuration['title'] ?? "$rate%");
 
         return $adjuster;
     }
 
     public function calculate(?object $subject = null, ?array $configuration = null): DetailedAmount
     {
-        $rate = floatval($configuration['rate'] ?? 0);
+        if (!$subject instanceof Adjustable) {
+            return new DetailedAmountDto(0);
+        }
 
-        return DetailedAmountDto::fromArray([['title' => "$rate%", 'amount' => $subject->preAdjustmentTotal() * $rate / 100]]);
+        $adjuster = $this->getAdjuster($configuration);
+
+        return DetailedAmountDto::fromArray([[
+            'title' => $adjuster->getTitle(),
+            'amount' => $adjuster->createAdjustment($subject)->getAmount(),
+        ]]);
     }
 
     public function getSchema(): Schema
     {
-        return Expect::structure(['rate' => Expect::float(0)->required()]);
+        return Expect::structure([
+            'rate' => Expect::float(0)->required(),
+            'title' => Expect::string()->nullable(),
+            'included' => Expect::bool()->nullable(),
+        ]);
     }
 
     public function getSchemaSample(array $mergeWith = null): array
     {
-        return ['rate' => 19];
+        return [
+            'rate' => 19,
+            'title' => 'Tax',
+            'included' => false,
+        ];
     }
 }