+++ src/Taxes/Contracts/TaxCalculator.php
@@ -24,10 +24,14 @@ interface TaxCalculator extends Schematized
     /**
      * We don't set the return type on the language level
      * since the adjustments module is optional
+     * @todo Remove this nonsense in v6 and add the proper return type
      *
      * @return null|\Vanilo\Adjustments\Contracts\Adjuster
      */
     public function getAdjuster(?array $configuration = null): ?object;
 
+    /**
+     * @todo change ?object to Adjustable in v6. Also - when can the configuration be null?
+     */
     public function calculate(?object $subject = null, ?array $configuration = null): DetailedAmount;
 }